Separate out ZKCoordination and SegmentManager
